requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-yasg>=1.21
python-dotenv==1.2.1
dj-database-url>=2.1
psycopg2-binary
google-auth>=2.23
requests>=2.31
gunicorn
whitenoise
resend>=2.0.0
